pyproject.toml
[project]
name = "essay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "codebasetwo", email = "codebasetwo@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.37.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-cli>=0.0.36",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.2",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "openai>=1.95.1",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.3",
    "tavily-python>=0.7.9",
]

[project.scripts]
essay = "essay:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff]
# Enable basic checks and import sorting
lint.select = ["E", "F", "I"]
extend-exclude = ["*.ipynb"]
line-length = 88  
# Exclude directories
exclude = [

    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "*.ipynb",
    "__pycache__",
]
# Target Python version
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
# Ignore unused imports in __init__.py files
"__init__.py" = ["F401"]
"gui.py" = ["F401"]
# Ignore missing type annotations in tests
"test_*.py" = ["ANN"]